+++ src/Aws/Support/Resources/support-2013-04-15.php
@@ -31,6 +31,75 @@ return array (
         ),
     ),
     'operations' => array(
+        'AddAttachmentsToSet' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\JsonCommand',
+            'responseClass' => 'AddAttachmentsToSetResponse',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Content-Type' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'application/x-amz-json-1.1',
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/json',
+                ),
+                'X-Amz-Target' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'AWSSupport_20130415.AddAttachmentsToSet',
+                ),
+                'attachmentSetId' => array(
+                    'type' => 'string',
+                    'location' => 'json',
+                ),
+                'attachments' => array(
+                    'required' => true,
+                    'type' => 'array',
+                    'location' => 'json',
+                    'items' => array(
+                        'name' => 'Attachment',
+                        'type' => 'object',
+                        'properties' => array(
+                            'fileName' => array(
+                                'type' => 'string',
+                            ),
+                            'data' => array(
+                                'type' => 'string',
+                                'filters' => array(
+                                    'base64_encode',
+                                ),
+                            ),
+                        ),
+                    ),
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'An internal server error occurred.',
+                    'class' => 'InternalServerErrorException',
+                ),
+                array(
+                    'reason' => 'An attachment set with the specified ID could not be found.',
+                    'class' => 'AttachmentSetIdNotFoundException',
+                ),
+                array(
+                    'reason' => 'The expiration time of the attachment set has passed. The set expires 1 hour after it is created.',
+                    'class' => 'AttachmentSetExpiredException',
+                ),
+                array(
+                    'reason' => 'A limit for the size of an attachment set has been exceeded. The limits are 3 attachments and 5 MB per attachment.',
+                    'class' => 'AttachmentSetSizeLimitExceededException',
+                ),
+                array(
+                    'reason' => 'The limit for the number of attachment sets created in a short period of time has been exceeded.',
+                    'class' => 'AttachmentLimitExceededException',
+                ),
+            ),
+        ),
         'AddCommunicationToCase' => array(
             'httpMethod' => 'POST',
             'uri' => '/',
@@ -72,6 +141,10 @@ return array (
                         'type' => 'string',
                     ),
                 ),
+                'attachmentSetId' => array(
+                    'type' => 'string',
+                    'location' => 'json',
+                ),
             ),
             'errorResponses' => array(
                 array(
@@ -81,6 +154,14 @@ return array (
                 array(
                     'reason' => 'The requested CaseId could not be located.',
                     'class' => 'CaseIdNotFoundException',
+                ),
+                array(
+                    'reason' => 'An attachment set with the specified ID could not be found.',
+                    'class' => 'AttachmentSetIdNotFoundException',
+                ),
+                array(
+                    'reason' => 'The expiration time of the attachment set has passed. The set expires 1 hour after it is created.',
+                    'class' => 'AttachmentSetExpiredException',
                 ),
             ),
         ),
@@ -146,6 +227,10 @@ return array (
                     'type' => 'string',
                     'location' => 'json',
                 ),
+                'attachmentSetId' => array(
+                    'type' => 'string',
+                    'location' => 'json',
+                ),
             ),
             'errorResponses' => array(
                 array(
@@ -155,6 +240,56 @@ return array (
                 array(
                     'reason' => 'The case creation limit for the account has been exceeded.',
                     'class' => 'CaseCreationLimitExceededException',
+                ),
+                array(
+                    'reason' => 'An attachment set with the specified ID could not be found.',
+                    'class' => 'AttachmentSetIdNotFoundException',
+                ),
+                array(
+                    'reason' => 'The expiration time of the attachment set has passed. The set expires 1 hour after it is created.',
+                    'class' => 'AttachmentSetExpiredException',
+                ),
+            ),
+        ),
+        'DescribeAttachment' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\JsonCommand',
+            'responseClass' => 'DescribeAttachmentResponse',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Content-Type' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'application/x-amz-json-1.1',
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/json',
+                ),
+                'X-Amz-Target' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'AWSSupport_20130415.DescribeAttachment',
+                ),
+                'attachmentId' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'json',
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'An internal server error occurred.',
+                    'class' => 'InternalServerErrorException',
+                ),
+                array(
+                    'reason' => 'The limit for the number of DescribeAttachment requests in a short period of time has been exceeded.',
+                    'class' => 'DescribeAttachmentLimitExceededException',
+                ),
+                array(
+                    'reason' => 'An attachment with the specified ID could not be found.',
+                    'class' => 'AttachmentIdNotFoundException',
                 ),
             ),
         ),
@@ -217,6 +352,11 @@ return array (
                 ),
                 'language' => array(
                     'type' => 'string',
+                    'location' => 'json',
+                ),
+                'includeCommunications' => array(
+                    'type' => 'boolean',
+                    'format' => 'boolean-string',
                     'location' => 'json',
                 ),
             ),
@@ -583,6 +723,20 @@ return array (
         ),
     ),
     'models' => array(
+        'AddAttachmentsToSetResponse' => array(
+            'type' => 'object',
+            'additionalProperties' => true,
+            'properties' => array(
+                'attachmentSetId' => array(
+                    'type' => 'string',
+                    'location' => 'json',
+                ),
+                'expiryTime' => array(
+                    'type' => 'string',
+                    'location' => 'json',
+                ),
+            ),
+        ),
         'AddCommunicationToCaseResponse' => array(
             'type' => 'object',
             'additionalProperties' => true,
@@ -600,6 +754,27 @@ return array (
                 'caseId' => array(
                     'type' => 'string',
                     'location' => 'json',
+                ),
+            ),
+        ),
+        'DescribeAttachmentResponse' => array(
+            'type' => 'object',
+            'additionalProperties' => true,
+            'properties' => array(
+                'attachment' => array(
+                    'type' => 'object',
+                    'location' => 'json',
+                    'properties' => array(
+                        'fileName' => array(
+                            'type' => 'string',
+                        ),
+                        'data' => array(
+                            'type' => 'string',
+                            'filters' => array(
+                                'base64_decode',
+                            ),
+                        ),
+                    ),
                 ),
             ),
         ),
@@ -662,6 +837,21 @@ return array (
                                                 'timeCreated' => array(
                                                     'type' => 'string',
                                                 ),
+                                                'attachmentSet' => array(
+                                                    'type' => 'array',
+                                                    'items' => array(
+                                                        'name' => 'AttachmentDetails',
+                                                        'type' => 'object',
+                                                        'properties' => array(
+                                                            'attachmentId' => array(
+                                                                'type' => 'string',
+                                                            ),
+                                                            'fileName' => array(
+                                                                'type' => 'string',
+                                                            ),
+                                                        ),
+                                                    ),
+                                                ),
                                             ),
                                         ),
                                     ),
@@ -711,6 +901,21 @@ return array (
                             ),
                             'timeCreated' => array(
                                 'type' => 'string',
+                            ),
+                            'attachmentSet' => array(
+                                'type' => 'array',
+                                'items' => array(
+                                    'name' => 'AttachmentDetails',
+                                    'type' => 'object',
+                                    'properties' => array(
+                                        'attachmentId' => array(
+                                            'type' => 'string',
+                                        ),
+                                        'fileName' => array(
+                                            'type' => 'string',
+                                        ),
+                                    ),
+                                ),
                             ),
                         ),
                     ),